[SSO][m]: controllers -> blueprints
Use blueprints as CKAN 2.8 uses flask
