+++ ckanext/ozwillo_pyoidc/oidc.py
@@ -91,7 +91,7 @@ class Client(oic.Client):
                 atresp = self.do_access_token_request(
                     scope="openid", state=authresp["state"], request_args=args,
                     authn_method=self.registration_response["token_endpoint_auth_method"])
-                id_token = atresp['id_token']
+                id_token = atresp.get('id_token', {})
                 app_admin = 'app_admin' in id_token and id_token['app_admin']
                 app_user = 'app_user' in id_token  and id_token['app_user']
             except Exception as err:
